# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Herald.lol Gaming Analytics API
  description: |
    # Herald.lol Gaming Analytics Platform API
    
    The Herald.lol API provides comprehensive access to League of Legends and Teamfight Tactics analytics, 
    player data, team management, and gaming insights. Built for developers, analysts, and gaming applications.
    
    ## Features
    
    - **Gaming Analytics**: KDA analysis, CS/min tracking, Vision Score, Damage Share, Gold Efficiency
    - **Riot API Integration**: Seamless access to Riot Games data with rate limit management
    - **Team Management**: Team creation, player management, performance tracking
    - **Real-time Data**: Live match analysis and gaming statistics
    - **Multi-region Support**: Global LoL regions (NA, EUW, KR, JP, etc.)
    
    ## Performance Targets
    
    - **Analysis Speed**: <5 seconds for gaming analytics
    - **UI Response**: <2 seconds for dashboard loading
    - **Uptime**: 99.9% platform availability
    - **Scalability**: Support for 1M+ concurrent users
    
    ## Authentication
    
    Herald.lol uses multiple authentication methods:
    
    - **Bearer Tokens**: JWT tokens for user authentication
    - **API Keys**: For server-to-server integration
    - **OAuth 2.0**: Social login integration
    - **MFA**: Multi-factor authentication for sensitive operations
    
    ## Rate Limiting
    
    Rate limits are enforced per subscription tier:
    
    - **Free**: 60 requests/minute
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Gaming-API-Key
      description: API key for server-to-server authentication
    
    OAuth2:
      type: oauth2
      description: OAuth2 authentication with multiple providers
      flows:
        authorizationCode:
          authorizationUrl: https://api.herald.lol/auth/oauth/authorize
          tokenUrl: https://api.herald.lol/auth/oauth/token
          scopes:
            read: Read access to gaming data
            write: Write access to gaming data
            admin: Administrative access
            analytics: Access to advanced analytics

  # Common schemas
  schemas:
...
    Error:
      type: object
      required:
        - error
        - gaming_platform
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        code:
          type: string
          description: Error code
        gaming_platform:
          type: string
          enum: [herald-lol]
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier

    # Gaming user schema
    GamingUser:
      type: object
      required:
        - id
        - username
        - gaming_platform
      properties:
        id:
          type: string
          description: Unique user identifier
        username:
          type: string
          description: Gaming username
        display_name:
          type: string
          description: Display name
        email:
          type: string
          format: email
        gaming_profile:
          $ref: '#/components/schemas/GamingProfile'
        subscription_tier:
          type: string
          enum: [free, premium, pro, enterprise]
        created_at:
          type: string
          format: date-time
        last_active:
          type: string
          format: date-time
        gaming_platform:
          type: string
          enum: [herald-lol]

    # Gaming profile schema
    GamingProfile:
      type: object
      properties:
        summoner_name:
          type: string
          description: League of Legends summoner name
        summoner_level:
          type: integer
          description: Summoner level
        region:
          type: string
          enum: [NA, EUW, EUNE, KR, JP, BR, LAN, LAS, OCE, RU, TR]
          description: Gaming region
        tier:
          type: string
          enum: [IRON, BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, MASTER, GRANDMASTER, CHALLENGER]
        rank:
          type: string
          enum: [IV, III, II, I]
        league_points:
          type: integer
          description: Current LP
        main_role:
          type: string
          enum: [top, jungle, mid, adc, support]
        favorite_champions:
          type: array
          items:
            type: string
        total_games:
          type: integer
        win_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1

    # Gaming statistics schema
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid gaming parameters provided"
            details: "Summoner name must be between 3-16 characters"
            gaming_platform: "herald-lol"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Gaming API authentication required"
            details: "Please provide a valid Bearer token or API key"
            gaming_platform: "herald-lol"
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                remember_me:
                  type: boolean
                  default: false
            example:
              email: "gamer@herald.lol"
              password: "securePassword123"
              remember_me: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
                  expires_in:
                    type: integer
                    description: Token expiration in seconds
                  user:
                    $ref: '#/components/schemas/GamingUser'
                  gaming_platform:
                    type: string
                    enum: [herald-lol]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register new gaming account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - summoner_name
                - region
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                summoner_name:
                  type: string
                  minLength: 3
                  maxLength: 16
                region:
                  type: string
                  enum: [NA, EUW, EUNE, KR, JP, BR, LAN, LAS, OCE, RU, TR]
            example:
              username: "herald_gamer"
              email: "gamer@herald.lol"
              password: "securePassword123"
              summoner_name: "HeraldChampion"
              region: "NA"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/GamingUser'
                  gaming_platform:
                    type: string
                    enum: [herald-lol]
        '400':
          $ref: '#/components/responses/BadRequest'

  # Gaming Analytics Endpoints
...